Use database for storing/fetching config files in lieu of RWX storage (#51)
* Store config files in the database

Rather than relying on a shared volume, store generated configuration
files in the database.

* Allow xdmod to write to /etc/httpd and /etc/xmod

The xmod user requires write access to these directories in order to
update the xdmod configuration and the Apache configuration at
runtime.

* Add GetConfigFiles.py

GetConfigFiles.py fetches configuration files from the
database and writes them to the local filesystem.

* Add xdmod-run-ingestor script

xdmod-run-ingestor is a script to fetch configuration files from the
database and then run the ingestor.

* Remove carriage returns from xdmod-get-config-files

* Prefer numeric file modes

The 'X' doesn't not seem to set the execute bit on the script, and I
am more use to seeing 774 as opposed to the symbolic modes.

* Upload tarball of /etc/xdmod directory to database

Rather than uploading/downloading individual files, tar up /etc/xdmod
and upload the resulting tarball to the database.

* Sleep before calling ingestor

We have been experiencing some database connectivity problems. This
commit attempts to reduce the rate of connections to the database by
adding a sleep before calling the ingestor.

* Simplify SQL error handling in GetConfigFiles.py

Fail with default traceback rather than trying to trap exceptions and
emit a custom error message.

* Simplify check for xdmod_init.json

Having both the os.path.isfile guard and the exception handling was
redundant. We can handle a missing file via an exception.

* Remove unused code from GetConfigFiles.py

* Add some docstrings to GetConfigFiles.py

* Shorten error message when xdmod-init.json is missing

* Use context manager form of subprocess.run

Using subprocess.run as a context manager ensures that we will
wait() on the child process.

* Simplify exception handling in xdmod-init.py

Fail with default traceback rather than trying to trap exceptions and
emit a custom error message.

* Remove unused code in xdmod_init.py

* Disable even more pylint warnings

* Clean up tarball generation

Run tar in verbose mode and clean up artifacts after uploading
the tarball to the database.

Co-authored-by: rob-baron <robbaron@bu.edu>