Fixed this deployment
Eliminated the comment indicating how obtaining an application token
was done, I am under the impression that it is done differently now.

Also eliminated the references to the openshift-xdmod-data volume as
it should not be mounted here.  This volume should be mounted on the job
and cannot be mounted here as we do not have RWX pvs.  By mounting it
here will prevent the job from starting.