Reorganization to support deploying with ArgoCD (#64)
* Permissions Changes

Chages were made to allow access to files in /etc/xdmod and /etc/httpd as
with the current permissions the following error appears when you try
to list the files:

ls -la /etc/xdmod
ls: cannot access /etc/xdmod/.: Permission denied
ls: cannot access /etc/xdmod/..: Permission denied
ls: cannot access /etc/xdmod/assets.json: Permission denied
ls: cannot access /etc/xdmod/colors1.json: Permission denied
ls: cannot access /etc/xdmod/datawarehouse.d: Permission denied
...
ls: cannot access /etc/xdmod/roles.d: Permission denied
ls: cannot access /etc/xdmod/roles.json: Permission denied
ls: cannot access /etc/xdmod/setup.json: Permission denied
ls: cannot access /etc/xdmod/update_check.json: Permission denied
ls: cannot access /etc/xdmod/user_management.json: Permission denied
total 0
d????????? ? ? ? ?            ? .
d????????? ? ? ? ?            ? ..
-????????? ? ? ? ?            ? assets.json
-????????? ? ? ? ?            ? colors1.json
d????????? ? ? ? ?            ? datawarehouse.d
-????????? ? ? ? ?            ? datawarehouse.json
...
d????????? ? ? ? ?            ? roles.d
-????????? ? ? ? ?            ? roles.json
-????????? ? ? ? ?            ? setup.json
-????????? ? ? ? ?            ? update_check.json
-????????? ? ? ? ?            ? user_management.json

This effectively prevest the process to copy what is currently in
those files on to the persistent volume that we will load containing
those files so that we can have processes write to those files and
not loose configuration when the pod is restarted.

* Removal oif specific branch

In the case of the Argo deploy, the specific branch will be specified
by the Argo overlay.

For a development deploy, the specific branch will be specified by th
development overlay

removed the branch as the branch will be overwirtten by the

* Only build the non-dev containers

For the Argo depoly, we do not want to build the dev containers
So these will have to built either manually or by the development
overlay.

* Fix some yamllint issues

This just fixes some issues like too many blank lines, or whitespace at the end
of the line.

* Organization to deploy on the Nerc using ArgoCD

This overlay combines the build and deploy yaml so that it is given
a single point of entry for ArgoCD to build and deploy this project
on to the NERC.

* Changing the Project name

Change the project name in the internal image registry to the generic
xdmod, which is the namespace that this is usually deployed under.

* Made naming convention more similar to an other project

As opposed to naming the overlay NERC, I reanmed it to be similar
to the valut project calling it nerc-ocp-infra which is more descriptive
of where it is deployed.

* Added back the dev builds

Found that the dev build has enough file permissions to copy over the
directories that is required by in the xdmod-init-1 init container.

The xdmod-init-2 initcontainer and the xdmod container can use the
regular container.

* Fix syntax in kustomization.yaml

Changed bases to resources.  Had to update my oc client, as resources
didn't work initially, but after updating it did.

* Fixed this deployment

Eliminated the comment indicating how obtaining an application token
was done, I am under the impression that it is done differently now.

Also eliminated the references to the openshift-xdmod-data volume as
it should not be mounted here.  This volume should be mounted on the job
and cannot be mounted here as we do not have RWX pvs.  By mounting it
here will prevent the job from starting.

* Remove the reference to a specific branch/commit

The specific branch/commit ti be build should be in the overlay
directories in order to control which branch/commit is built in either
development, staging or production.