Support for Volume events (#58)
* Now supporting volume events.

* did a quick clean up of pylint errors I was ignoring

* Addressing code review comments

Added a "if os.path.isfile(...)" check before the "with open( ... ) as ...":
there are times that the file might not exist.  No sense in throwing an error
that isn't really an error.

Set the initialized the last_run_date time to the epoch time.  This way is
isn't None.

* Which initialization to keep

I had intialized one variable twice, in part to changing how
I was thinking about the use of that variable and function
it was in.  I kepts the events_by_date={} as this both fit the
more recent way of doing things.