Need to include the yaml to add the namespace for Argo CD (#68)
* Need to include the yaml to add the namespace for Argo CD

I had figured that argo cd created namespaces from the namespace
defined in other files as I was able to find the xdmod project in the
openshift UI.  However, this incorrect.

Adding this in her as Justin found the following error:

    Failed sync attempt to 5c594eeb996faf241c9ce67d7967b3428fb8a4e8:
    one or more objects failed to apply,
    reason: namespaces "xdmod" not found (retried 5 times).

* Useing IncludeNamespace as a proxy for a specific commit

When this finally gets deployed using Argo CD, will change these
to a specific commit.  Until then, will use the branch name as a proxy.