Fix invalid build config (#69)
* Fixed a build config error

This error appeared when I tried to apply the build config as that had
previously been defined:

"k8s/overlays/nerc-ocp-infra": BuildConfig.build.openshift.io "bc-xdmod-dev" is invalid: spec.strategy.dockerStrategy.from.name: Required value

I had to replace it with what the container has in the from.  Behavioraly,
kustomize went from defaulting to what is in the from of the docker file
to requiring the from container in the build config.  I liked the older
behavior better than then new one.

* Use FixInvalidBc as the proxy for the commit to use in production

Use FixInvlidBc as the production version as the problems that this
fixed in ocp-staging are likely to occur in production.

* Update bc-xdmod-dev.yaml

Added new-line as per @naved001 review comment