Run the python scripts to fetch data from openstack in python 3.9 container
Decided to run the python scripts with a current version of python.  This
needs to be done as the connection to mariadb/mysql will no longer
support the version that is available on centos 7, and xdmod will be moving
to something newer in the future anyways.

This will run as an init container before the shredder.