Deploy Shred Job
These are the files to deploy the openstack shredding job.
1) the configuration is loaded from the database via the init container
2) The data from openstack is loaded into the PV in the init container
3) xdmod-shredder is run in the container.

This was done as we don't yet support RWX persistent volumes and the
xdmod team is still on centos 7 which uses an older version of of python
of which they are planning to update to a current version of python.

I would like to patch the args to the command in the overlay/nerc-ocp-infra
patch directory, but seem to be hitting https://github.com/kubernetes-sigs/kustomize/issues/4062