Rework of the Ingestor job (#121)
* Rework of the Ingestor job

Split off the downloading of the config from the database into
an init container from the running of xdmod-ingestor in the
main container.

* Stop shortening get-config

xdmod-get-config is a better name than xdmod-get-conf.

* Use the correct image

I don't see a reason why I should prepent a moc for images at this
time as they are potentially useful outside of the moc.

* corrected the script

One of the annoying differences between centos and the python3 container
is the location of the shell command.  This fixes it (pthyon 3: #!/bin/bash,
centos 7: #!/usr/bin/sh)

* Delete an unused file

Am no longer calling xdmod-ingestor from the shell script
as the shell script has been replaced by xdmod-get-configs.sh
which is called in a init container.

* Calling the Correct Command

On the python container all execuatable files are loaded into the
/app directory as opposed to on the centos based xdmod container where
they are installed in /usr/bin.