Fix Bug 730 take 2 (#133)
* Fix Bug 730 take 2

This bug appears twice, fixed the volume one while missing the
VM one.  We iterate through the timestamps looking at VMs then
volumee and eliminating the ones that are not updated.

* Fix bug 730

We want to ignore VMs (and volumes)  that are not updated or messages that come
from the are not handled. The last Fix bug 730 missed the line at 730
and I don't think it kept the same semantics.

* Fix Bug 730 take 2

Checking to see that the key exists via better python semantics:
<key> in <dict>

* Fixed a typo,

update should have been updated