Work-A-Round for Running xdmod's acl-config (#145)
* Work-A-Round for Running xdmod's acl-config

It seems that everytime the shredder run, the xdmod-ui interface gets
stuck in the loading process.  This is either due to acl-config
not running to completion on the shdredder pod, or acl-config modifying
the config directory and scince the config directory is not shared RWX
it doesn't updated on the xdmod-ui pod.

This will require more investigation as to what is occurng.  Meanwhile,
this will run acl-config in the xdmod-ui pod every 7 minutes.  I'm trying to
strike a balance between running acl-config too often (consumning too much
database time) and keeping the UI from being stuck loading.

* Elimination of the port and Improved Readability

Deleted the port from the work-a-round container to remove potential
contention.  Also took lars up on his suggestion to improve the
readbility of the bash code.\