Add image build workflow for Dockerfile.moc-xdmod
The image is built for both pull requests and pushes.

The image is pushed to ghcr.io for:

- Pushes to main, which will be tagged with the branch name and the
  commit id.

- Pushes for semver [1] style tags (`vX.Y.Z`), which will be tagged
  with `vX.Y.Z`, `vX.Y`, `vX`, and `latest`.

This ensures that the Dockerfile builds successfully, makes the image
available for deployment or local testing, and opens up the possibility
of implementing some CI tests of the image in future work.

If it weren't for the dependency between `moc-xdmod` and
`moc-xdmod-dev` we could use a simple `matrix` [2] build strategy to
build multiple images, but with the dependency we need a fair amount
of duplicate code in the workflow.

[1]: https://semver.org/
[2]: https://docs.github.com/en/actions/using-jobs/using-a-matrix-for-your-jobs