Hiearchy job (#157)
* inital commit

In the middle of integrating the data format of the coldfront
data export.

* Shortening functions

eliminating variables that are nolonger useful

* Ready for a draft PR

Have moved sensitive data to files external to this project

The sample data coming from coldfront is being processed

Yet to do:

1 write specific function to get the data from coldfront
2 write specific function to get the data from keycloak
3 process data from keycloak into the format required to generate files
4 rework the file generation process to simplify
5 add the names.csv to rename projects to the <coldfront-project> - <openstack/openshift project>

* Some changes based on code review
.
the most major change is not deleting the database everytime
when running the script

* changed based on review

* responding to review comments

working though changes made to the hierarchy ids and
simplifying the code.

* Continuing to work though the ids for the file

* Back to generating running to completion

finally straighted out the ids, and considerably simplified it.

* Am processing the keycloak data

Generating xdmod hierarchy files that look reasonable from the
combination the coldfront and keycloak data

* eliminate duplicate code

the code that was eliminated was made redunant by fixing
the id logic in the processing_data function

* Corrects the formatting of the pi2projects and names csv

Need to configure the xdmod-init file to specify the mapping
between xdmod resrouce and what is known in coldfront

* Updated xdmod_init.json

with an example link between coldfront and xdmod resrouce

* review comments

* required changes to fix the format of pi2projects.csv

Need to mapp the project to the resource.  Only cloud project
are included in this as there is not an equivalent file for
hpc.

* Blacked this file

I have made no changes to it, I am placing it here as it
is example code to fetch data from keyclock

* new commit message

* Added for so that pytest can start from a known state

Need to delete the database in order that pytest can
start from a known state.

* Eliminate unsed comments

Just eliminated some commented out code that was used for debugging.

* Added test_002 to start testing for the second run

Change the data that was used in test_001 a bit to remove
a PI and projects, and want to see that pi and those projects
removed from the output.

* Review comments

Added more testing and changed the format of PIs to
their name followed by their email address.

* Just keeping this file similar to the one in UpdateDocs

Trying to keep these the same in order that when that is updated
there are no changes needed to process_hierarchy.py

* Revew Comment

Added names to tests and comments about the dataset sets.

* Review Comment

Added names to each of the numbered tests.

* Pull the data from keycloak

This fetches the metadata that is stored in keycloak relating PIs,
the PI's field of science, and the PI's institution via keycloak's
API.

* Added connection, addressed review comments and updated tests

This turned in the a larger commit than expected.  Added the functions
to connect with coldfront and keycloak - still getting an error
from coldfront.

Updated the tests, and refactored some code to simplify the processing
and testing.

* Added a cli to specify the directory to write the files in

This will specify the diretory to place the files in from the command
line which will be specified in the comman to run from the cron job
yaml

* Removed email from PI's name

Use just the pi name as opposed to concatenating the pi's name with the
pi's email address.

* Updated tests

updated the tests to specify the correct directory to use and
to only use the pi's name (no mention of their email address).

* Added reference to process_hierarchy.py

Added the process_hierarchy.py script to the dockerfil

* Kubnetes manifests to deploy the cj-xdmod-hierarchy

These are the manifests to deploy the hierarchy job.
    1) create a pvc for the pv to store the hierarchy files
    2) create the job
        2.1) init container - to download the config
        2.2) init container - using the keycloak creds (in config),
             fetch the coldfront and keycloak data
        2.3) container - process the heirarchy files
    3) create patches for deployment on staging vs production

* Changed names for consistency

Changed the names to be consistent with the rest of the files
in the kubernetes manifests.

* Fixed the staging overlays

The staging overlays lagged behine the production overlays, This bring
them up to date.


* Active mongo (#181)

Added the mongo deployment and svc to allow argo cd to deploy
on the nerc

* Moved time of xdmod openshift cron job (#182)

The job currently runs at 1am; however 1am server time is apparently
9pm the previous day OpenShift time. Changing the time to 1pm will
allow the job to get all the data for the previous day.

* Increae of maraidb storage (#184)

Increase the the size of the PV that is mounted to mariadb's data
directory.

* refactored the tests

This is responding to a suggestion from a review comment.

* Reviewing data from openshift:
1) some combined coldfront - openshift project names were too long for xdmod
2) There is some confusion as to what project name gets inserted in to the group name
   - wanted to handle if the cold front project name is intereted into the group name

* Read the Config from xdmod_init.json

Get the database passwords from xdmod_init.json as opposed to
hard coding them.

* Add in Section for ColdFront

The cold front info section is required to run the
process_hierarchy as it contains the REST endpoint for
ColdFront

* As of mariadb 10.4 (and backported to 10.3) Differences in MariaDB

MariaDB has a different way of handling the root account at least for
an install on a local system (doesn't seem to affect the container that
is running mariadb).  This handles the way a privileged account needs
to access it via the unix socket.

* Handle differences between running tests via command line and vscode

had to reset the directories so that pytest could successively run
tests in the container

* Concatenated coldfront and project names are sometimes too long

In the data setion, had to eliminage the concatenated
ColdFront Project Name - (OpenStack | OpenShift Name) as these
names became too long in xdmod.

* Docker file that can build an image to run the test

The built image contains the tests and all of the dependences to
run the tests.

* The initialization script for mariadb

This initializes mariadb in terms of basic database setup and
an xdmod account.  The xdmod account is there for completeness, though
it is currently not used

* The xdmod_init.json Configuration File

This is a minimal xdmod_init.json configuration file to run
these tests.

* Update README.md

Now has a testing section that includes how to build the container
and how to start the database and run the tests.

* Added instructions to run tests

New section added to explain how to run the tests for processing
the hierarchy.

* Changed the way that testing was being done.

Non-existent branch referenced.  Was being used for testing, but
testing was done a different way.

---------

Co-authored-by: Tzu-Mainn Chen <tzumainn@redhat.com>